app now needs separate tests with this upgraded version of angular-cli
