+++ src/app/tsconfig.app.json
@@ -26,7 +26,7 @@
     "flatModuleId": "ngx-cbp-theme"
   },
   "files": [
-    "./index.ts",
+    "./test-entry.ts",
     "./typings.d.ts"
   ]
