+++ src/app/tsconfig.spec.json
@@ -16,7 +16,8 @@
       "node_modules/@types"
     ],
     "files": [
-      "test-entry.ts", "test.ts"
+      "./test-entry.ts",
+      "./test.ts"
     ],
     "include": [
       "**/*.spec.ts",
